refactor: 命名規則の修正 Fixes #77
